+++ src-tauri/src/main.rs
@@ -10,6 +10,7 @@ use tauri::Manager;
 
 #[tokio::main]
 async fn main() {
+    #[cfg(target_os = "linux")]
     std::env::set_var("WEBKIT_DISABLE_DMABUF_RENDERER", "1");
     env_logger::init();