+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# [1.1.0-beta.2](https://github.com/franklinnolasco7/Vivant/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2026-07-11)
+
+
+### Bug Fixes
+
+* support Windows EPUB images and annotation sorting ([8b4d8f9](https://github.com/franklinnolasco7/Vivant/commit/8b4d8f90bef2783e20cff16d21733e501626e888))
+
+# [1.1.0-beta.1](https://github.com/franklinnolasco7/Vivant/compare/v1.0.0...v1.1.0-beta.1) (2026-07-11)
+
+
+### Bug Fixes
+
+* isolate version sync from CLI entrypoint ([54b1e3b](https://github.com/franklinnolasco7/Vivant/commit/54b1e3b9585e91894011ffb38d61aaaa7c44a9cf))
+
+
+### Features
+
+* add Windows prerelease support ([2ccfd37](https://github.com/franklinnolasco7/Vivant/commit/2ccfd379a4e107714967886575fb3f5dad3af343))
+
 # 1.0.0 (2026-07-11)